refactor the routes for flask
